+++ ecom-backend/src/main/java/com/ecommerce/project/payload/OrderItemDTO.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,8 +15,8 @@ public class OrderItemDTO {
     private Long orderItemId;
     private ProductDTO product;
     private Integer quantity;
-    private double discount;
-    private double orderedProductPrice;
+    private BigDecimal discount;
+    private BigDecimal orderedProductPrice;
 
 
 }
